try to consolidate entrypoint and modal function models
