+++ garden-backend-service/src/api/schemas/entrypoint.py
@@ -3,61 +3,24 @@ from uuid import UUID
 from pydantic import AliasPath, Field
 
 from .base import BaseSchema, UniqueList, Url
+from .shared_function_schemas import CommonFunctionMetadata, CommonFunctionPatchRequest
 
 
-class _RepositoryMetadata(BaseSchema):
-    repo_name: str
-    url: Url
-    contributors: UniqueList[str] = Field(default_factory=list)
-
-
-class _PaperMetadata(BaseSchema):
-    title: str
-    authors: UniqueList[str] = Field(default_factory=list)
-    doi: str | None
-    citation: str | None
-
-
-class _DatasetMetadata(BaseSchema):
-    title: str = Field(...)
-    doi: str | None
-    url: Url
-    data_type: str | None
-    repository: str
-
-
-# protected_namespaces=() to allow model_* attribute names
-class _ModelMetadata(BaseSchema, protected_namespaces=()):
-    model_identifier: str
-    model_repository: str
-    model_version: str | None
-
-
-class EntrypointMetadata(BaseSchema):
+class EntrypointMetadata(CommonFunctionMetadata):
+    # Entrypoints always have a DOI
+    # They can be in draft or published state
     doi: str
     doi_is_draft: bool
-    title: str
-    description: str | None
-    year: str
+
+    # Metadata specific to Globus Compute functions
     func_uuid: UUID
     container_uuid: UUID
     base_image_uri: str
     full_image_uri: str
     notebook_url: Url
-    is_archived: bool = False
 
-    short_name: str
-    function_text: str
-
-    authors: UniqueList[str] = Field(default_factory=list)
-    tags: UniqueList[str] = Field(default_factory=list)
-    test_functions: list[str] = Field(default_factory=list)
-    requirements: list[str] = Field(default_factory=list)
-
-    models: list[_ModelMetadata] = Field(default_factory=list)
-    repositories: list[_RepositoryMetadata] = Field(default_factory=list)
-    papers: list[_PaperMetadata] = Field(default_factory=list)
-    datasets: list[_DatasetMetadata] = Field(default_factory=list)
+    # The function name is stored as "short_name" for entrypoints
+    short_name: str | None = None
 
 
 class EntrypointCreateRequest(EntrypointMetadata):
@@ -69,27 +32,11 @@ class EntrypointMetadataResponse(EntrypointMetadata):
     id: int
 
 
-class EntrypointPatchRequest(BaseSchema):
+class EntrypointPatchRequest(CommonFunctionPatchRequest):
     doi_is_draft: bool | None = None
-    title: str | None = None
-    description: str | None = None
-    year: str | None = None
+    
     func_uuid: UUID | None = None
     container_uuid: UUID | None = None
     base_image_uri: str | None = None
     full_image_uri: str | None = None
     notebook_url: Url | None = None
-    is_archived: bool | None = None
-
-    short_name: str | None = None
-    function_text: str | None = None
-
-    authors: UniqueList[str] | None = None
-    tags: UniqueList[str] | None = None
-    test_functions: list[str] | None = None
-    requirements: list[str] | None = None
-
-    models: list[_ModelMetadata] | None = None
-    repositories: list[_RepositoryMetadata] | None = None
-    papers: list[_PaperMetadata] | None = None
-    datasets: list[_DatasetMetadata] | None = None
